SCRUM-64 rename: 결제 전체조회 마스터 서비스 분리
